+++ .github/workflows/run_tests.yaml
@@ -22,5 +22,3 @@ jobs:
         uses: actions/checkout@v2
       - name: Test
         run: go test ./...
-      - name: Test cmd execution
-        run: sudo go run ./cmd/hidra/main.go --test --file ./examples
