+++ Makefile
@@ -2,15 +2,15 @@ BUILD_PATH = 'build'
 VERSION = $(shell git describe --tags --abbrev=0)
 
 versionize:
-	sed -i 's/latest/${VERSION}/g' utils/version.go
+	sed -i 's/latest/${VERSION}/g' src/utils/version.go
 
 build:
-	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-amd64/hidra cmd/hidra/main.go
-	CGO_ENABLED=0 GOOS=linux GOARCH=386 go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-386/hidra cmd/hidra/main.go
-	CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-arm64/hidra cmd/hidra/main.go
-	CGO_ENABLED=0 GOOS=linux GOARCH=arm go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-arm/hidra cmd/hidra/main.go
-	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -o ${BUILD_PATH}/hidra-${VERSION}-darwin-amd64/hidra cmd/hidra/main.go
-	CGO_ENABLED=0 GOOS=darwin GOARCH=arm64 go build -o ${BUILD_PATH}/hidra-${VERSION}-darwin-arm64/hidra cmd/hidra/main.go
+	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-amd64/hidra src/cmd/hidra/main.go
+	CGO_ENABLED=0 GOOS=linux GOARCH=386 go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-386/hidra src/cmd/hidra/main.go
+	CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-arm64/hidra src/cmd/hidra/main.go
+	CGO_ENABLED=0 GOOS=linux GOARCH=arm go build -o ${BUILD_PATH}/hidra-${VERSION}-linux-arm/hidra src/cmd/hidra/main.go
+	CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -o ${BUILD_PATH}/hidra-${VERSION}-darwin-amd64/hidra src/cmd/hidra/main.go
+	CGO_ENABLED=0 GOOS=darwin GOARCH=arm64 go build -o ${BUILD_PATH}/hidra-${VERSION}-darwin-arm64/hidra src/cmd/hidra/main.go
 deps:
 	go mod vendor
 
@@ -25,7 +25,7 @@ compress:
 
 clean:
 	rm -rf build
-	sed -i 's/${VERSION}/latest/g' utils/version.go
+	sed -i 's/${VERSION}/latest/g' src/utils/version.go
 
 
 all: deps build compress
