+++ src/scenarios/browser/browser.go
@@ -6,10 +6,9 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/hidracloud/hidra/models"
-	"github.com/hidracloud/hidra/scenarios"
-
 	"github.com/chromedp/chromedp"
+	"github.com/hidracloud/hidra/src/models"
+	"github.com/hidracloud/hidra/src/scenarios"
 )
 
 // Scenario Represent an browser scenario
