+++ src/scenarios/tls/tls.go
@@ -8,9 +8,9 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/hidracloud/hidra/models"
-	"github.com/hidracloud/hidra/scenarios"
-	"github.com/hidracloud/hidra/utils"
+	"github.com/hidracloud/hidra/src/models"
+	"github.com/hidracloud/hidra/src/scenarios"
+	"github.com/hidracloud/hidra/src/utils"
 )
 
 // Scenario Represent an ssl scenario
